feat: frame tag and rendering rules rework
New rendering rules discern tags into two types: inline rendered and block rendered.
Inline tags are alweys renderd next to each other within the same line, if there are
any white-spaces betweeb the inline elements it will be replaced with a single space.
Block tags, start with a line break character, unless the block element is the first
one and end with a line break, unless the block element is the last one.